% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_subtract.R
\name{background_subtract}
\alias{background_subtract}
\title{Background correction for chromatograms.}
\usage{
background_subtract(chrom, back)
}
\arguments{
\item{chrom}{Data frame with the chromatographic data.}

\item{back}{Data frame with the background data.}
}
\value{
Data frame of the background-subtracted sample chromatogram.
}
\description{
This function takes background data from a sample run and subtracts it from
a sample chromatogram. It is ONLY intended for use following the extraction
of chromatographic data using \code{read_wacl_ch}, as it relies on the data frame
structure from the \code{read_wacl_ch} function.
}
\author{
Thomas Warburton
}
