% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mad.R
\name{compute_mad}
\alias{compute_mad}
\title{Compute Median Absolute Deviation (MAD)}
\usage{
compute_mad(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing the MAD.
}
\description{
Compute Median Absolute Deviation (MAD)
}
\author{
Thomas Warburton
}
