% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mod_z.R
\name{compute_mod_z}
\alias{compute_mod_z}
\title{Compute Modified Z-Scores, and add columns to the end of existing dataframe}
\usage{
compute_mod_z(df)
}
\arguments{
\item{df}{A data frame with numeric columns.}
}
\value{
A data frame with modified Z-scores added as new columns.
}
\description{
Compute Modified Z-Scores, and add columns to the end of existing dataframe
}
