% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_detection.R
\name{detect_noise}
\alias{detect_noise}
\title{Detect Noise Regions and Manage Chromatographic Data}
\usage{
detect_noise(
  df,
  rt_col_name = NULL,
  intensity_col_name = NULL,
  smoothing_window = 10,
  noise_threshold = 5,
  min_noise_duration = 5,
  keep_noise = TRUE
)
}
\arguments{
\item{df}{A data frame containing chromatographic data with retention time
and intensity columns. The function will attempt to identify the columns
automatically based on typical names if not specified by the user.}

\item{rt_col_name}{Optional. A string specifying the column name for retention time.
If not provided, the function will attempt to detect it automatically.}

\item{intensity_col_name}{Optional. A string specifying the column name for intensity.
If not provided, the function will attempt to detect it automatically.}

\item{smoothing_window}{An integer specifying the window size for smoothing
the intensity data. Default is 10. Reducing this size increases sensitivity.}

\item{noise_threshold}{A numeric value specifying the threshold below which
the smoothed intensity is considered noise. Default is 5. Lowering this value
makes the function more sensitive.}

\item{min_noise_duration}{Minimum number of consecutive points to consider
as a noise region. Default is 5. Reducing this value helps in detecting shorter
noise bursts.}

\item{keep_noise}{Logical. If \code{TRUE}, the function replaces non-noise values
with \code{NA}. If \code{FALSE}, the function removes noise regions from the data frame
entirely and keeps only non-noise regions. Default is \code{TRUE}.}
}
\value{
A data frame with the same structure as the input \code{df}, but with
either non-noise values or noise values replaced with \code{NA} based
on the \code{keep_noise} parameter.
}
\description{
For the identification of areas of noise in chromatographic data based on
retention time and intensity. It can either keep the noise regions (and
replace non-noise regions with \code{NA}) or remove the noise regions (and keep
non-noise regions). As with any type of noise detection, there are obvious
possibilities non-noise data will be retained if you wish to isolate noise-
data only, and vice versa. As such you will need to adjust the parameters
accordingly, on the understanding you will likely need to compromise in
order to isolate either noise or non-noise. If you wish to run calculations
on the noise data, you will need to apply rolling windows on your data.
Running \code{detect_noise} twice could be needed for data with large peaks.
}
\examples{
df <- data.frame(rt = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
                 signal = c(2, 3, 2, 2, 10, 50, 2, 2, 2, 2))
detect_noise(df, smoothing_window = 3, noise_threshold = 1, min_noise_duration = 2, keep_noise = TRUE)
detect_noise(df, keep_noise = FALSE)

}
\author{
Thomas Warburton
}
