% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilution_volumes.R
\name{dilution_volumes}
\alias{dilution_volumes}
\title{Calculate the volume required to dilute a calibration mix to a required final
fraction.}
\usage{
dilution_volumes(
  cylinder_volume,
  cal_fraction,
  fill_pressure,
  target_fraction,
  ambient_pressure = 1
)
}
\arguments{
\item{cylinder_volume}{The volume of the cylinder the dilutions will be done in.
Must be a numeric value.}

\item{cal_fraction}{The fraction of the calibration mix being used for the
dilution. Must be a numeric value.}

\item{fill_pressure}{The pressure of the final target gas mix, in units of bar
(gauge). Must be a numeric value.}

\item{target_fraction}{The required fraction of the final target gas. Can be a
single numeric value or a vector of numeric values.}

\item{ambient_pressure}{The ambient pressure of the cylinder. Default value
of 1 bar.}
}
\value{
A data frame with the target_fraction, fill_pressure, cal_fraction,
and the calculated dilution_volume.
}
\description{
This should be used in the context of creating a target gas at a specified
mixing fraction (ppb/ppm). The function will return a volume (or vector of
volumes if a vector of fractions was given) required in the cylinder volume
specified diluted to the specified final pressure. Use common units between
the two fraction parameters. Please use 'bar' as the pressure units.
}
\examples{
library(iaaR)

fractions=c(0.1, 0.2, 0.3, 0.4, 0.5)

df <- dilution_volumes(cylinder_volume=45, 
                       fill_pressure=100,
                       cal_fraction=4,
                       target_fraction=fractions)


}
\author{
Thomas Warburton
}
