% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_to_velocity.R
\name{f2v}
\alias{f2v}
\title{Convert volumetric flow rate (L / min) to velocity (m / s)}
\usage{
f2v(flow_rate, pipe_diameter)
}
\arguments{
\item{flow_rate}{A numeric giving the flow rate of the fluid in units of
litres per minute (L / min)}

\item{pipe_diameter}{A numeric giving the diamteter of the pipe in units
of metres (m)}
}
\value{
A numeric of the calculated velocity of the fluid.
}
\description{
A function to convert the flow rate \code{Q} (L / min) of a fluid in a pipe of
diameter \code{D} (m) into a velocity \code{V} (m / s). Make sure you use the correct
units!
}
\examples{
library(iaaR)

result <- f2v(90, 0.037)
print(result)
}
\author{
Thomas Warburton
}
