% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment.R
\name{increment}
\alias{increment}
\title{Calculate the concentration increment from a steady emission rate.}
\usage{
increment(emission_rate, volume, aer, volume_step = NULL, aer_step = NULL)
}
\arguments{
\item{emission_rate}{The emission rate of the VOC in grams per hour (g/hr)}

\item{volume}{A numeric of the room volume in metres cubed. Can be a vector
of lower and upper volumes, in which case a volume step of 0.1 will be used
to produce the sequence.}

\item{aer}{A numeric of the air exchange rate (AER) per hour. Can be a vector
of lower and upper AER, in which case an AER step of 0.01 will be used to
produce the sequence.}

\item{volume_step}{A numeric used to indicate the steps between volume
values when creating the volume sequence. Only to be used if using a vector
input to \code{volume}.}

\item{aer_step}{A numeric used to indicate the steps between AER
values when creating the AER sequence. Only to be used if using a vector
input to \code{aer}.}
}
\value{
A data frame with the calculated VOC concentration increment, as well
as the volume and AER inputs. VOC increment is in micrograms per cubic metre.
}
\description{
This function calculates the concentration increment from a steady emission
rate. This is calculated according to a steady-state assumption, using a
one-box model. This is only for an upper-bound estimation, as this model does
not account for chemical loss through deposition/surface sinking or through
oxidation reactions. Chemical loss in this model is only through ventilation.
The intended use of this function is to calculate an in-room VOC increment
through a point emission.
}
\examples{
library(iaaR)

df <- increment(emission_rate = 0.005, 
volume = 25, 
aer = 0.75)

print(df)

library(iaaR)

df_expanded <- increment(emission_rate = 0.005, 
volume = c(5,25), 
aer = c(0.1,0.75))

print(df_expanded)

}
\author{
Thomas Warburton
}
