% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_limits.R
\name{instrument_limits}
\alias{instrument_limits}
\title{Calculate the limits of detection (LOD) or quantification/quantitation (LOQ). This will only work with a data frame with VOC names as col heads, and each sample as a separate row. Limits calculated through dividing the linearity residuals of a calibration curve against the slope of the regression. Values within the data frame should be GC peak response (eg. peak area)}
\usage{
instrument_limits(
  df,
  limit_method,
  concs = NULL,
  concs_in_df = FALSE,
  concs_units = NULL,
  zero_remove = TRUE,
  conc_col_position = "none"
)
}
\arguments{
\item{df}{A data frame with numeric columns}

\item{limit_method}{"detection", "quant" or "both". Calculates either LOD, LOQ or does both. Must be specified.}

\item{concs}{Creates a value of the concentrations used in the calibration curve. Must be specified if 'concs_in_df' is FALSE.}

\item{concs_in_df}{Logical to return a TRUE or FALSE for whether the concentrations are included in df. FALSE by default.}

\item{zero_remove}{Logical to change any 0 into an NA in df. TRUE by default.}

\item{concs_col_position}{State the position of the concentration column in df. Must be either "first" or "last" if 'concs_in_df' is TRUE. "none" if 'concs_in_df" is FALSE. "none" by default.}
}
\value{
A data frame with the LOD/LOQ calculated for each column in df
}
\description{
Calculate the limits of detection (LOD) or quantification/quantitation (LOQ). This will only work with a data frame with VOC names as col heads, and each sample as a separate row. Limits calculated through dividing the linearity residuals of a calibration curve against the slope of the regression. Values within the data frame should be GC peak response (eg. peak area)
}
