% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_complete.R
\name{keep_complete}
\alias{keep_complete}
\title{Remove all columns with a single NA or 0 value}
\usage{
keep_complete(df)
}
\arguments{
\item{df}{The data frame to have columns removed from}
}
\value{
The data frame originating from \code{df}, with the specified columns
returned
}
\description{
A function to tidy a data frame to keep only columns with a complete series
of observations. Will remove any columns which contain at least one NA or
0 value. Treats NaN and Inf as NA.
}
\author{
Thomas Warburton
}
