% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquid_injection.R
\name{liquid_injection}
\alias{liquid_injection}
\title{Calculate the volume of liquid required to generate a target gas mixture}
\usage{
liquid_injection(name, mixing_ratio, volume, mw = NULL, density = NULL)
}
\arguments{
\item{name}{The name or CAS number of the chemical used to generate the
target gas.}

\item{mixing_ratio}{The mixing ratio of the target gas, expressed in parts
per billion (ppb).}

\item{volume}{The total volume of gas being blended, in cubic metres (m^3).}

\item{mw}{An optional numeric value specifying the molecular weight of the
chemical (in g/mol). If omitted, the function attempts to retrieve it
automatically.}

\item{density}{An optional numeric value specifying the density of the
liquid chemical (in g/cm^3). If omitted, the function attempts to retrieve
it automatically.}
}
\value{
A data frame containing the chemical name, mixing ratio, gas volume,
molecular weight, density, and the calculated liquid injection volume in
microlitres.
}
\description{
This function calculates the volume of a liquid chemical required to produce
a target gas mixture at a specified mixing ratio within a vessel. It assumes
that the liquid fully evaporates into the gas phase and does not partition
or adsorb within the vessel.
}
\details{
Provide the function with the chemical name (or CAS number), the desired
mixing ratio of the target gas (in parts per billion), and the final gas
volume within the vessel (in metres cubed). The final gas volume can be
calculated, for example, by multiplying the vessel volume by the final
pressure (in bar) and then adding the physical vessel volume (ie. a 200 bar,
10 litre vessel will have a final volume of 2010 L, or 2.01 m^3).

The function will attempt to automatically retrieve the chemical's density
and molecular weight from PubChem. However, either or both of these values
may be supplied directly via the \code{density} and \code{mw} arguments if preferred.
}
\examples{
library(iaaR)

df <- liquid_injection("ethanol", 4, 2.01)
view(df)

df_CAS_density_mw_provided <- liquid_injection(
  "64-17-5", 4, 2.01,
  density = 0.79,
  mw = 46.07
)
view(df_CAS_density_mw_provided)


}
\author{
Thomas Warburton
}
