% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rows.R
\name{make_rows}
\alias{make_rows}
\title{Elongate a dataframe by adding NA rows at the end}
\usage{
make_rows(df, number_of_total_rows)
}
\arguments{
\item{df}{The data frame to be made longer.}

\item{number_of_total_rows}{A numeric with the number of rows to extend \code{df}
up to.}
}
\value{
\code{df} with additional NA rows added to be made equal to
\code{number_of_total_rows}.
}
\description{
This function takes a data frame and adds in a set number of NA rows. Mainly
for creating data frames of equal length.
}
\author{
Thomas Warburton
}
