% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcm_mass.R
\name{mcm_mass}
\alias{mcm_mass}
\title{Extract molecular mass from Master Chemical Mechanism (MCM)}
\usage{
mcm_mass(species_name)
}
\arguments{
\item{species_name}{A character variable for the name of the primary
pollutant to extract molecular mass for. MUST follow the same naming
mechanism as used in the MCM.}
}
\value{
A numeric of the scraped molecular mass for the named species.
}
\description{
The MCM is a near-explicit mechanism for the degradation of atmospheric
pollutants to the final oxidation products of carbon dioxide and water.
This function will extract the molecular mass for the named species from
the MCM website.
}
\author{
Thomas Warburton
}
