% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCM_rates.R
\name{MCM_rates}
\alias{MCM_rates}
\title{Extract reaction rates from Master Chemical Mechanism (MCM)}
\usage{
MCM_rates(species_name)
}
\arguments{
\item{species_name}{A character variable for the name of the primary
pollutant to extract reaction rates for. MUST follow the same naming
mechanism as used in the MCM.}
}
\value{
A data frame with the names of each oxidant, the rate of reaction
between the primary species with the oxidant, and the oxidation product.
}
\description{
The MCM is a near-explicit mechanism for the degradation of atmospheric
pollutants to the final oxidation products of carbon dioxide and water.
This function will extract the primary reaction rates for the reaction of
the gas-phase reactant with an oxidant. This function does not include
precursor reaction rates (rates for the formation of the species in
question). Essentially, this function is a data scraper.
}
\author{
Thomas Warburton
}
