% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wacl_ch.R
\name{read_wacl_ch}
\alias{read_wacl_ch}
\title{Read and Parse Agilent ChemStation .ch Files}
\usage{
read_wacl_ch(path, format_out = "data.frame", data_format = "long")
}
\arguments{
\item{path}{Path to the \code{.ch} file.}

\item{format_out}{Format of the output data: either "matrix" or "data.frame". Default is "data.frame".}

\item{data_format}{Format of the data: "wide" for a single column of intensity values,
or "long" for separate columns for retention time and intensity. Default is "long".}
}
\value{
A dataframe or matrix of retention time and peak intensity in the specified format.
}
\description{
This function reads Agilent \code{.ch} files and extracts chromatogram data
such as intensity and retention times. This is a heavily edited version of
code originally from chromConverter. See note for more information. This
function does not scale intensities extracted from \code{.ch} file. For use
primarily for users of the LABGCFIDQMS instrument at Wolfson Atmospheric
Chemistry Laboratories (WACL)
}
\note{
This function is derived from a script originally written by Ethan Bass.
Please visit github.com/ethanbass/chromConverter to see the original script,
plus many more functions for reading chromatography files into R. I (TW) only
adapted the code, and take no credit whatsoever for it. Please see the reference
below for Ethan's original code, called by \code{read_chemstation_ch}, plus many more
functions for reading chromatograms into R in. This version of \code{read_chemstation_ch}
exists solely to read .ch files into a format with retention time and intensity values.
}
\references{
Bass, E. (2023). chromConverter: Chromatographic File Converter. http://doi.org/10.5281/zenodo.6792521.
}
\author{
Thomas Warburton and Ethan Bass

#' @export
}
