% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reynolds_number.R
\name{reynolds_number}
\alias{reynolds_number}
\title{Calculates Reynold's number for flow in a circular tube}
\usage{
reynolds_number()
}
\arguments{
\item{None.}{The function prompts for user input interactively in the
console.}
}
\value{
A numeric of the calculated Reynold's number.
}
\description{
This function calculates the Reynold's number for a fluid flowing through a
circular tube. The function prompts for all necessary parameters
interactively in the console.
}
\examples{
result <- reynolds_number()
print(result)

}
\author{
Thomas Warburton
}
