% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube_length.R
\name{tube_length}
\alias{tube_length}
\title{Calculate the length of tube for a required volume}
\usage{
tube_length(id, vol, units)
}
\arguments{
\item{id}{A numeric giving the internal diameter (ID) of the tube being used.}

\item{vol}{A numeric giving the required internal volume.}
}
\value{
A numeric of the calculated tube length.
}
\description{
This function is intended to be used to calculate the length of tube/pipe
required for a set internal volume. Generally for the creation of a measured
target gas, or to create a gas spiking volume.
}
\author{
Thomas Warburton
}
