% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_to_flow.R
\name{v2f}
\alias{v2f}
\title{Convert fluid velocity (m / s) to volumetric flow rate (L / min)}
\usage{
v2f(velocity, pipe_diameter)
}
\arguments{
\item{velocity}{A numeric giving the velocity of the fluid in units of
metres per second (m / s)}

\item{pipe_diameter}{A numeric giving the diamteter of the pipe in units
of metres (m)}
}
\value{
A numeric of the calculated flow rate of the fluid in units of
litres per minute (L / min).
}
\description{
A function to convert the flow rate \code{V} (m / s) of a fluid in a pipe of
diameter \code{D} (m) into a velocity \code{Q} (L / min). Make sure you use the
correct units!
}
\examples{
library(iaaR)

result <- v2f(15, 0.037)
print(result)
}
\author{
Thomas Warburton
}
