% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_filter.R
\name{z_filter}
\alias{z_filter}
\title{Replace values with NA based on Modified Z-Scores}
\usage{
z_filter(
  df,
  z_threshold_upper = 3.5,
  z_threshold_lower = -3.5,
  replace_zeros = TRUE
)
}
\arguments{
\item{df}{A data frame with numeric columns.}

\item{replace_zeros}{Logical indicating whether to replace zeros with NA. Default is TRUE.}

\item{z_upper}{Numeric value for upper threshold. Default is 3.5.}

\item{z_lower}{Numeric value for lower threshold. Default is -3.5.}
}
\value{
A data frame with values replaced based on thresholds.
}
\description{
Replace values with NA based on Modified Z-Scores
}
