% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_diff.R
\name{date_diff}
\alias{date_diff}
\title{Find the difference between two dates}
\usage{
date_diff(initial_date, final_date, date_format = "\%Y-\%m-\%d")
}
\arguments{
\item{initial_date}{The first date to be considered, generally the 'start' date
of the time scale. Must be same format as 'final_date'.}

\item{final_date}{The final date to be considered, generally the 'end' date
of the time scale. Must be same format as 'initial_date'.}

\item{date_format}{The date format. Must be identical across 'initial_date' and
'final_date'. Use Y, M and D in quotes to express the date
format (e.g "\%Y-\%m-\%d" or "\%d-\%m-\%Y"). Taken as
"YYYY-MM-DD" as default. Can also use}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         A numeric value giving the difference between the two dates.
}\if{html}{\out{</div>}}
}
\description{
This function will find the different between two dates in number of days.
This only find the day difference, and does not account for times, that is
each date is taken as \verb{YYYY-DD-MM 00:00:00}.
}
\examples{
initial_date <- "2023-01-01"
final__date <- "2023-04-24"
result <- date_diff(initial_date, final_date, date_format="\%Y-\%m-\%d")
print(result)

}
\author{
Thomas Warburton
}
