% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_detection.R
\name{detect_noise}
\alias{detect_noise}
\title{Detect Noise Regions and Manage Chromatographic Data}
\usage{
detect_noise(
  df,
  rt_col_name = NULL,
  intensity_col_name = NULL,
  smoothing_window = 5,
  noise_threshold = 2,
  min_noise_duration = 3,
  keep_noise = TRUE
)
}
\arguments{
\item{df}{A data frame containing chromatographic data with retention time
and intensity columns. The function will attempt to identify the columns
automatically based on typical names if not specified by the user.}

\item{rt_col_name}{Optional. A string specifying the column name for retention time.
If not provided, the function will attempt to detect it automatically.}

\item{intensity_col_name}{Optional. A string specifying the column name for intensity.
If not provided, the function will attempt to detect it automatically.}

\item{smoothing_window}{An integer specifying the window size for smoothing
the intensity data. Default is 5. Reducing this size increases sensitivity.}

\item{noise_threshold}{A numeric value specifying the threshold below which
the smoothed intensity is considered noise. Default is 2. Lowering this value
makes the function more sensitive.}

\item{min_noise_duration}{Minimum number of consecutive points to consider
as a noise region. Default is 3. Reducing this value helps in detecting shorter
noise bursts.}

\item{keep_noise}{Logical. If \code{TRUE}, the function replaces non-noise values
with \code{NA}. If \code{FALSE}, the function removes noise regions from the data frame
entirely and keeps only non-noise regions. Default is \code{TRUE}.}
}
\value{
A data frame with the same structure as the input \code{df}, but with
either non-noise values or noise values replaced with \code{NA} based
on the \code{keep_noise} parameter.
}
\description{
This function identifies areas of noise in chromatographic data based on retention
time and intensity. It offers an option to either keep the noise regions (and
replace non-noise regions with \code{NA}) or remove the noise regions (and keep
non-noise regions). Non-noise regions can be replaced with \code{NA} or removed
from the data frame depending on the specified option.
}
\examples{
df <- data.frame(rt = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
                 signal = c(2, 3, 2, 2, 10, 50, 2, 2, 2, 2))
detect_noise(df, smoothing_window = 3, noise_threshold = 1, min_noise_duration = 2, keep_noise = TRUE)
detect_noise(df, smoothing_window = 3, noise_threshold = 1, min_noise_duration = 2, keep_noise = FALSE)
}
