% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_limits.R
\name{instrument_limits}
\alias{instrument_limits}
\title{Calculate the limits of detection (LOD) or quantification/quantitation (LOQ).}
\usage{
instrument_limits(
  df,
  limit_method,
  concs = NULL,
  concs_in_df = FALSE,
  zero_remove = TRUE,
  conc_col_position = "none",
  df_concs = NULL
)
}
\arguments{
\item{df}{A data frame with numeric columns representing peak areas.}

\item{limit_method}{Character string specifying the type of limit to calculate:
"detection", "quant" or "both".}

\item{concs}{Optional vector of concentrations used in the calibration curve.
Must be provided if 'concs_in_df' is FALSE.}

\item{concs_in_df}{Logical indicating whether concentrations are included in df.}

\item{zero_remove}{Logical indicating whether to replace zeros with NA in df.}

\item{df_concs}{Optional data frame with concentrations corresponding to each species
in df. Must be of equal size to df.}

\item{concs_col_position}{Position of the concentration column in df if 'concs_in_df'
is TRUE. Must be either "first", "last", or "none".}
}
\value{
A data frame with the LOD/LOQ calculated for each column in df.
}
\description{
This function calculates the LOD or LOQ for each VOC (volatile organic compound)
using the peak area as input data. Limits are calculated by dividing the linearity
residuals of a calibration curve against the slope of the regression. Regression
Pearson's R. Doesn't allow for Spearman's Rho / Kendall's Tau (these  methods
are more for either non-parametric regressions or those with outliers, and if your
calibration curve contains outliers, it probably needs to be run again).
}
