% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcm_rates.R
\name{mcm_rates}
\alias{mcm_rates}
\title{Extract reaction rates from Master Chemical Mechanism (MCM)}
\usage{
mcm_rates(species_name)
}
\arguments{
\item{species_name}{A character variable for the name of the primary
pollutant to extract reaction rates for. MUST follow the same naming
mechanism as used in the MCM.}
}
\value{
A data frame with the names of each oxidant, the rate of reaction
between the primary species with the oxidant stored as a string, and the
oxidation product.
}
\description{
The MCM is a near-explicit mechanism for the degradation of atmospheric
pollutants to the final oxidation products of carbon dioxide and water.
This function will extract the primary reaction rates for the reaction of
the gas-phase reactant with an oxidant. This function does not include
precursor reaction rates (rates for the formation of the species in
question). Essentially, this function is a data scraper. However, this is
only intended for primary VOCs, as with secondary VOCs with several precursor
reactions, the rates often turn very complex - in this case, you are probably
better off downloading a FACSIMILE (.fac) file directly from the website
(mcm.york.ac.uk). This function will not attempt to resolve any rates which
make reference to complex rates or photolysis parameters - in this case, you
will be better served referring directly to the MCM.
}
\examples{
#This is example code for how to use the `mcm_rates` function to scrape reaction
rates for alpha-pinene, and then use the scraped rates mathematically in R.

library(iaaR)

df <- mcm_rates("APINENE") 

df_rates <- df["Rate"]

T <- 293

df_results <- data.frame(Result = sapply(df_rates$Rate, function(expr) eval(parse(text = expr))))

}
\author{
Thomas Warburton
}
