% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure_drop.R
\name{pressure_drop}
\alias{pressure_drop}
\title{Calculates pressure drop in circular tubing}
\usage{
pressure_drop()
}
\arguments{
\item{None.}{The function prompts for user input interactively in the
console.}
}
\value{
A list containing the following:
\itemize{
\item \code{flow_type}: The calculated flow type ('laminar' or 'turbulent').
\item \code{flow_orientation}: Either 'horizontal' or 'vertical'.
\item \code{flow_direction}: Whether vertical flow was 'upwards' or 'downwards'.
Only applicable for 'vertical' flow.
\item \code{reynolds_number}: The calculated Reynolds number.
\item \code{pressure_drop}: The calculated pressure drop in Pascals.
}
}
\description{
This function calculates the pressure drop in a circular tube for either
laminar or turbulent flow. There is currently no logic for if a calculated
Reynolds number indicates the flow is in a laminar-turbulent transition
state (taken as having a Reynolds number between 2300 and 4000), but the
function does display warnings for Reynolds numbers that indicate this
transition. Laminar flow is calculated using the Hagen-Poiseuille equation,
while turbulent flow is calculated using the Darcy-Weisbach equation. Now
with added vertical functionality, calculating vertical pressure gains or
losses through calculation of wall shear stressed for both laminar and
turbulent flow. The function prompts for all necessary parameters
interactively in the console.
}
\examples{
result <- pressure_drop()
print(result$pressure_drop)

}
\author{
Thomas Warburton
}
