% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ch.R
\name{read_ch}
\alias{read_ch}
\title{Read and Parse Agilent ChemStation .ch Files}
\usage{
read_ch(
  path,
  format_out = c("data.frame", "matrix", "df", "data", "mat", "m"),
  data_format = c("long", "wide")
)
}
\arguments{
\item{path}{Path to the \code{.ch} file.}

\item{format_out}{Format of the output data: either "matrix" or "data.frame". Default is "data.frame".}

\item{data_format}{Format of the data: "wide" for a single column of intensity values,
or "long" for separate columns for retention time and intensity. Default is "long".}
}
\value{
A chromatogram in the specified format.
}
\description{
This function reads Agilent \code{.ch} files and extracts chromatogram data
such as intensity and retention times. It supports various file formats from
different versions of Agilent ChemStation and OpenLab software.
}
\note{
This function is derived from a script originally written by Ethan Bass.
Please visit github.com/ethanbass/chromConverter to see the original script,
plus many more functions for reading chromatography files into R. I (TW) only
adapted the code, and take no credit whatsoever for it. Please see the reference
below for Ethan's original code, called by \code{read_chemstation_ch}, plus many more
functions for reading chromatograms into R in. This version of \code{read_chemstation_ch}
exists solely to read .ch files into a format with retention time and intensity values.
}
\references{
Bass, E. (2023). chromConverter: Chromatographic File Converter. http://doi.org/10.5281/zenodo.6792521.
}
\author{
Thomas Warburton and Ethan Bass
}
