% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_filter.R
\name{z_filter}
\alias{z_filter}
\title{Replace values with NA based on Modified Z-Scores}
\usage{
z_filter(df, z_upper = 3.5, z_lower = -3.5, return_tidy = TRUE)
}
\arguments{
\item{df}{A data frame with numeric columns.}

\item{z_upper}{Numeric value for the upper threshold of modified Z-scores. Default is 3.5.}

\item{z_lower}{Numeric value for the lower threshold of modified Z-scores. Default is -3.5.}

\item{return_tidy}{Logical indicating whether to remove modified Z-score columns from the returned data frame. Default is TRUE.}
}
\value{
A data frame with outliers replaced by NA, optionally excluding modified Z-score columns.
}
\description{
This function replaces values in a data frame with NA based on thresholds
of modified Z-scores. It is important to ensure that filtering is done
ethically and only when justified.
}
